requirements.txt
bidict==0.21.2
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-SocketIO==4.3.1
Flask-WTF==0.14.3
flatbuffers==1.12
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
Werkzeug==1.0.1
WTForms==2.3.3
